{{cookiecutter.directory_name}}/requirements.txt
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==2.1.1
envyaml==1.10.211231
fastapi==0.75.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
pyaml==21.10.1
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.17.1
text-unidecode==1.3
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
